+++ packages/frontend/src/module/common/component/input/XrpAmountTextField/XrpAmountTextField.styles.ts
@@ -2,7 +2,7 @@ import styled, { css } from "styled-components";
 import NumericTextField from "../NumericTextField/NumericTextField";
 
 export const XrpAmountTextFieldRoot = styled(NumericTextField)(
-    ({ theme }) => css`
+    () => css`
         .Label {
             width: 100%;
         }
