+++ packages/frontend/src/module/settings/hook/useSettings.ts
@@ -1,0 +1,6 @@
+import { useRecoilValue } from "recoil";
+import { settingsState } from "../SettingsState";
+
+export default () => {
+    return useRecoilValue(settingsState);
+};
