+++ packages/frontend/test/src/module/common/component/feedback/ActionModal/ActionModal.spec.tsx
@@ -1,8 +1,8 @@
 import { screen } from "@testing-library/react";
 import { render, translate } from "test-utils";
 import userEvent from "@testing-library/user-event";
-import ActionModal from "module/common/component/feedback/DeprecatedActionModal/ActionModal";
-import { ActionFn } from "module/common/component/feedback/ActionModal/ActionModal.types";
+import ActionModal from "module/common/component/feedback/ActionModal/ActionModal";
+import { ActionFn } from "module/common/component/feedback/TabsModal/ActionModal.types";
 
 describe("ActionModal", () => {
     test("Renders correctly", () => {
